Task 2 upd

+++ JS Aplicatinos/05 JS-Applications-Asyncronous-Programming-Exercise-Resources/02. FISHER-GAME/app.js
@@ -6,19 +6,29 @@ const html = {
 }
 const loadHdr = () => fetchData('catches',{method: "GET"},[printData]);
 const getUrl = (x) => `https://fisher-game.firebaseio.com/${x}.json`;
-
+const clearField = () => html.cathesCtn().innerHTML ='<legend>Catches</legend>';
 function attachEvents() {
    html.loadBtn().addEventListener('click',loadHdr);
    html.addForm().querySelector('button').addEventListener('click',addHdr);
 }
 function fetchData(url,headers,callbacks){
     fetch(getUrl(url),headers)
-    .then(res=>res.json())
-    .then((data) => callbacks.forEach(x=>x(data)))
-    .catch((e) =>console.log(e));
+    .then((res)=>{
+		if (res.ok !== true) throw new Error('Cannot load resourse!');
+		return res.json();
+	})
+    .then((data) => {
+		if(data === null && headers.method === 'GET')
+		{
+			clearField();
+			throw new Error('Database is empty!');
+		}
+		callbacks.forEach(x=>x(data))
+		})
+    .catch((e) =>alert(e));
 }
-function printData(data){    
-    html.cathesCtn().innerHTML = '<legend>Catches</legend>';
+function printData(data){
+	clearField();
     Object.keys(data).forEach(id=>{
         const div = getEl('div',
         [{k:'class',v:'catch'},{k:'data-id',v:id}],
@@ -26,14 +36,16 @@ function printData(data){
     const updBtn = getEl('button',[{k:'class',v:'update'}],'Update');
     updBtn.addEventListener('click',updHandler.bind({id}));   
     const delBtn = getEl('button',[{k:'class',v:'update'}],'Delete');
-    delBtn.addEventListener('click',delHandler.bind({id})); 
+    delBtn.addEventListener('click',delHandler.bind({id,})); 
     div.appendChild(updBtn);
     div.appendChild(delBtn);
     html.cathesCtn().appendChild(div);
     });  
 }
 function delHandler(){
-    fetchData(`catches/${this.id}`,{method: "DELETE"},[loadHdr]);
+	 if(confirm(`Do you want to delete catch with id ${this.id}?`)){
+		fetchData(`catches/${this.id}`,{method: "DELETE"},[loadHdr]);
+	 }
 }
 function addHdr(){
     const obj = getDataObj(html.addForm().querySelectorAll('input'));
@@ -44,9 +56,7 @@ function addHdr(){
 function updHandler(){
     const obj = getDataObj(document.querySelector(`[data-id="${this.id}"]`)
                             .querySelectorAll('input'));
-    fetchData(`catches/${this.id}`,
-    {method: 'PUT',body: JSON.stringify(obj)},
-    [loadHdr]);
+    fetchData(`catches/${this.id}`,{method: 'PUT',body: JSON.stringify(obj)},[loadHdr]);
 }
 function getDataObj(x){
     return {
